+++ src/Field.php
@@ -15,26 +15,21 @@ use craft\elements\Category;
 use craft\elements\Entry;
 use craft\helpers\ArrayHelper;
 use craft\helpers\Db;
-use craft\helpers\FileHelper;
 use craft\helpers\Html;
-use craft\helpers\HtmlPurifier;
 use craft\helpers\Json;
-use craft\helpers\StringHelper;
-use craft\helpers\UrlHelper;
+use craft\htmlfield\events\ModifyPurifierConfigEvent;
+use craft\htmlfield\HtmlField;
+use craft\htmlfield\HtmlFieldData;
 use craft\models\Section;
-use craft\models\Site;
 use craft\redactor\assets\field\FieldAsset;
 use craft\redactor\assets\redactor\RedactorAsset;
-use craft\redactor\events\ModifyPurifierConfigEvent;
 use craft\redactor\events\ModifyRedactorConfigEvent;
 use craft\redactor\events\RegisterLinkOptionsEvent;
 use craft\redactor\events\RegisterPluginPathsEvent;
-use craft\validators\HandleValidator;
 use HTMLPurifier_Config;
 use yii\base\Event;
 use yii\base\InvalidArgumentException;
 use yii\base\InvalidConfigException;
-use yii\db\Schema;
 
 /**
  * Redactor field type
@@ -42,7 +37,7 @@ use yii\db\Schema;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 3.0
  */
-class Field extends \craft\base\Field
+class Field extends HtmlField
 {
     /**
      * @event RegisterPluginPathsEvent The event that is triggered when registering paths that contain Redactor plugins.
@@ -60,15 +55,18 @@ class Field extends \craft\base\Field
      * Plugins can get notified when HTML Purifier config is being constructed.
      *
      * ```php
-     * use craft\redactor\events\ModifyPurifierConfigEvent;
      * use craft\redactor\Field;
+     * use craft\htmlfield\ModifyPurifierConfigEvent;
      * use HTMLPurifier_AttrDef_Text;
      * use yii\base\Event;
      *
-     * Event::on(Field::class, Field::EVENT_MODIFY_PURIFIER_CONFIG, function(ModifyPurifierConfigEvent $e) {
-     *      // Allow the use of the Redactor Variables plugin
-     *      $e->config->getHTMLDefinition(true)->addAttribute('span', 'data-redactor-type', new HTMLPurifier_AttrDef_Text());
-     * });
+     * Event::on(
+     *     Field::class,
+     *     Field::EVENT_MODIFY_PURIFIER_CONFIG,
+     *     function(ModifyPurifierConfigEvent $event) {
+     *          // ...
+     *     }
+     * );
      * ```
      */
     const EVENT_MODIFY_PURIFIER_CONFIG = 'modifyPurifierConfig';
@@ -111,42 +109,6 @@ class Field extends \craft\base\Field
      * @var string|null The Redactor config file to use
      */
     public ?string $redactorConfig = null;
-
-    /**
-     * @var string|null The HTML Purifier config file to use
-     */
-    public ?string $purifierConfig = null;
-
-    /**
-     * @var bool Whether the HTML should be cleaned up on save
-     * @deprecated in 2.4
-     */
-    public bool $cleanupHtml = true;
-
-    /**
-     * @var bool Whether disallowed inline styles should be removed on save
-     */
-    public bool $removeInlineStyles = true;
-
-    /**
-     * @var bool Whether empty tags should be removed on save
-     */
-    public bool $removeEmptyTags = true;
-
-    /**
-     * @var bool Whether non-breaking spaces should be replaced by regular spaces on save
-     */
-    public bool $removeNbsp = true;
-
-    /**
-     * @var bool Whether the HTML should be purified on save
-     */
-    public bool $purifyHtml = true;
-
-    /**
-     * @var string The type of database column the field should have in the content table
-     */
-    public string $columnType = Schema::TYPE_TEXT;
 
     /**
      * @var string|array|null The volumes that should be available for Image selection.
@@ -254,6 +216,8 @@ class Field extends \craft\base\Field
             $config['showUnpermittedVolumes'] = true;
         }
 
+        unset($config['cleanupHtml']);
+
         parent::__construct($config);
     }
 
@@ -350,8 +314,8 @@ class Field extends \craft\base\Field
 
         return Craft::$app->getView()->renderTemplate('redactor/_field_settings', [
             'field' => $this,
-            'redactorConfigOptions' => $this->_getCustomConfigOptions('redactor'),
-            'purifierConfigOptions' => $this->_getCustomConfigOptions('htmlpurifier'),
+            'redactorConfigOptions' => $this->configOptions('redactor'),
+            'purifierConfigOptions' => $this->configOptions('htmlpurifier'),
             'volumeOptions' => $volumeOptions,
             'transformOptions' => $transformOptions,
             'defaultTransformOptions' => array_merge([
@@ -390,32 +354,9 @@ class Field extends \craft\base\Field
     /**
      * @inheritdoc
      */
-    public function getContentColumnType(): string
+    protected function createFieldData(string $content, ?int $siteId): HtmlFieldData
     {
-        return $this->columnType;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function normalizeValue($value, ElementInterface $element = null): ?FieldData
-    {
-        if ($value instanceof FieldData) {
-            return $value;
-        }
-
-        // Temporary fix (hopefully) for a Redactor bug where some HTML will get submitted when the field is blank,
-        // if any text was typed into the field, and then deleted
-        if ($value === '<p><br></p>') {
-            $value = null;
-        }
-
-        if (!$value) {
-            return null;
-        }
-
-        // Prevent everyone from having to use the |raw filter when outputting RTE content
-        return new FieldData($value, $element->siteId ?? null);
+        return new FieldData($content, $siteId);
     }
 
     /**
@@ -423,8 +364,6 @@ class Field extends \craft\base\Field
      */
     protected function inputHtml($value, ElementInterface $element = null): string
     {
-        /** @var FieldData|null $value */
-
         // register the asset/redactor bundles
         $view = Craft::$app->getView();
         $view->registerAssetBundle(FieldAsset::class);
@@ -484,14 +423,9 @@ class Field extends \craft\base\Field
         RedactorAsset::registerTranslations($view);
         $view->registerJs('new Craft.RedactorInput(' . Json::encode($settings) . ');');
 
-        if ($value instanceof FieldData) {
-            $value = $value->getRawContent();
-        }
+        $value = $this->prepValueForInput($value, $element);
 
-        if ($value !== null) {
-            // Parse reference tags
-            $value = $this->_parseRefs($value, $element);
-
+        if ($value !== '') {
             // Swap any <!--pagebreak-->'s with <hr>'s
             $value = str_replace('<!--pagebreak-->', Html::tag('hr', '', [
                 'class' => 'redactor_pagebreak',
@@ -519,40 +453,15 @@ class Field extends \craft\base\Field
      */
     public function getStaticHtml($value, ElementInterface $element): string
     {
-        /** @var FieldData|null $value */
-        return Html::tag('div', $value ?: '&nbsp;', [
-            'class' => array_filter([
-                'text',
-                $this->uiMode === 'enlarged' ? 'readable' : null,
-            ]),
-        ]);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function searchKeywords($value, ElementInterface $element): string
-    {
-        $keywords = parent::searchKeywords($value, $element);
-
-        if (Craft::$app->getDb()->getIsMysql()) {
-            $keywords = StringHelper::encodeMb4($keywords);
-        }
-
-        return $keywords;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function isValueEmpty($value, ElementInterface $element): bool
-    {
-        if ($value === null) {
-            return true;
-        }
-
-        /** @var FieldData $value */
-        return parent::isValueEmpty($value->getRawContent(), $element);
+        return
+            Html::beginTag('div', [
+                'class' => array_filter([
+                    'text',
+                    $this->uiMode === 'enlarged' ? 'readable' : null,
+                ]),
+            ]) .
+            ($this->prepValueForInput($value, $element) ?: '&nbsp;') .
+            Html::endTag('div');
     }
 
     /**
@@ -560,203 +469,16 @@ class Field extends \craft\base\Field
      */
     public function serializeValue($value, ElementInterface $element = null): ?string
     {
-        /** @var FieldData|null $value */
-        if (!$value) {
-            return null;
+        if ($value instanceof HtmlFieldData) {
+            $value = $value->getRawContent();
         }
 
-        // Get the raw value
-        $value = $value->getRawContent();
-
-        if (!$value) {
-            return null;
+        if (is_string($value)) {
+            // Swap any pagebreak <hr>'s with <!--pagebreak-->'s
+            $value = preg_replace('/<hr class="redactor_pagebreak".*?>/', '<!--pagebreak-->', $value);
         }
 
-        // Swap any pagebreak <hr>'s with <!--pagebreak-->'s
-        $value = preg_replace('/<hr class="redactor_pagebreak".*?>/', '<!--pagebreak-->', $value);
-
-        if ($this->purifyHtml) {
-            // Parse reference tags so HTMLPurifier doesn't encode the curly braces
-            $value = $this->_parseRefs($value, $element);
-
-            // Sanitize & tokenize any SVGs
-            $svgTokens = [];
-            $svgContent = [];
-            $value = preg_replace_callback('/<svg\b.*>.*<\/svg>/Uis', function(array $match) use (&$svgTokens, &$svgContent): string {
-                $svgContent[] = Html::sanitizeSvg($match[0]);
-                return $svgTokens[] = 'svg:' . StringHelper::randomString(10);
-            }, $value);
-
-            $value = HtmlPurifier::process($value, $this->_getPurifierConfig());
-
-            // Put the sanitized SVGs back
-            $value = str_replace($svgTokens, $svgContent, $value);
-        }
-
-        if ($this->removeInlineStyles) {
-            // Remove <font> tags
-            $value = preg_replace('/<(?:\/)?font\b[^>]*>/', '', $value);
-
-            // Remove disallowed inline styles
-            $allowedStyles = $this->_allowedStyles();
-            $value = preg_replace_callback(
-                '/(<(?:h1|h2|h3|h4|h5|h6|p|div|blockquote|pre|strong|em|b|i|u|a|span|img|table|thead|tbody|tr|td|th)\b[^>]*)\s+style="([^"]*)"/',
-                function(array $matches) use ($allowedStyles) {
-                    // Only allow certain styles through
-                    $allowed = [];
-                    $styles = explode(';', $matches[2]);
-                    foreach ($styles as $style) {
-                        [$name, $value] = array_map('trim', array_pad(explode(':', $style, 2), 2, ''));
-                        if (isset($allowedStyles[$name])) {
-                            $allowed[] = "$name: $value";
-                        }
-                    }
-                    return $matches[1] . (!empty($allowed) ? ' style="' . implode('; ', $allowed) . '"' : '');
-                },
-                $value
-            );
-        }
-
-        if ($this->removeEmptyTags) {
-            // Remove empty tags
-            $value = preg_replace('/<(h1|h2|h3|h4|h5|h6|p|div|blockquote|pre|strong|em|a|b|i|u|span)\s*><\/\1>/', '', $value);
-        }
-
-        if ($this->removeNbsp) {
-            // Replace non-breaking spaces with regular spaces
-            $value = preg_replace('/(&nbsp;|&#160;|\x{00A0})/u', ' ', $value);
-            $value = preg_replace('/  +/', ' ', $value);
-        }
-
-        // Find any element URLs and swap them with ref tags
-        $value = preg_replace_callback(
-            '/(href=|src=)([\'"])([^\'"\?#]*)(\?[^\'"\?#]+)?(#[^\'"\?#]+)?(?:#|%23)([\w\\\\]+)\:(\d+)(?:@(\d+))?(\:(?:transform\:)?' . HandleValidator::$handlePattern . ')?\2/',
-            function($matches) {
-                [, $attr, $q, $url, $query, $hash, $elementType, $ref, $siteId, $transform] = array_pad($matches, 10, null);
-
-                // Create the ref tag, and make sure :url is in there
-                $ref = $elementType . ':' . $ref . ($siteId ? "@$siteId" : '') . ($transform ?: ':url');
-
-                if ($query || $hash) {
-                    // Make sure that the query/hash isn't actually part of the parsed URL
-                    // - someone's Entry URL Format could include "?slug={slug}" or "#{slug}", etc.
-                    // - assets could include ?mtime=X&focal=none, etc.
-                    /** @noinspection PhpUnnecessaryCurlyVarSyntaxInspection */
-                    $parsed = Craft::$app->getElements()->parseRefs("{{$ref}}");
-                    if ($query) {
-                        // Decode any HTML entities, e.g. &amp;
-                        $query = Html::decode($query);
-                        if (mb_strpos($parsed, $query) !== false) {
-                            $url .= $query;
-                            $query = '';
-                        }
-                    }
-                    if ($hash && mb_strpos($parsed, $hash) !== false) {
-                        $url .= $hash;
-                        $hash = '';
-                    }
-                }
-
-                return $attr . $q . '{' . $ref . '||' . $url . '}' . $query . $hash . $q;
-            },
-            $value);
-
-        // Swap any regular URLS with element refs, too
-
-        // Get all URLs, sort by longest first.
-        $sortArray = [];
-        $siteUrlsById = [];
-        foreach (Craft::$app->getSites()->getAllSites(false) as $site) {
-            if ($site->hasUrls) {
-                $siteUrlsById[$site->id] = $site->getBaseUrl();
-                $sortArray[$site->id] = strlen($siteUrlsById[$site->id]);
-            }
-        }
-        arsort($sortArray);
-
-        $value = preg_replace_callback(
-            '/(href=|src=)([\'"])(http.*)?\2/',
-            function($matches) use ($sortArray, $siteUrlsById) {
-                $url = $matches[3] ?? null;
-
-                if (!$url) {
-                    return '';
-                }
-
-                // Longest URL first
-                foreach ($sortArray as $siteId => $bogus) {
-                    // Starts with a site URL
-
-                    if (StringHelper::startsWith($url, $siteUrlsById[$siteId])) {
-                        // Drop query
-                        $query = parse_url($url, PHP_URL_QUERY);
-
-                        if (!empty($query)) {
-                           break;
-                        }
-
-                        $uri = preg_replace('/\?.*/', '', $url);
-
-                        // Drop page trigger
-                        $pageTrigger = Craft::$app->getConfig()->getGeneral()->getPageTrigger();
-                        if (strpos($pageTrigger, '?') !== 0) {
-                            $pageTrigger = preg_quote($pageTrigger, '/');
-                            $uri = preg_replace("/^(?:(.*)\/)?$pageTrigger(\d+)$/", '', $uri);
-                        }
-
-                        // Drop site URL.
-                        $uri = StringHelper::removeLeft($uri, $siteUrlsById[$siteId]);
-
-                        if ($element = Craft::$app->getElements()->getElementByUri($uri, $siteId, true)) {
-                            $refHandle = $element::refHandle();
-                            $url = '{' . $refHandle . ':' . $element->id . '@' . $siteId . ':url||' . $url . '}';
-                            break;
-                        }
-                    }
-                }
-
-                return $matches[1] . $matches[2] . $url . $matches[2];
-            },
-            $value);
-
-        if (Craft::$app->getDb()->getIsMysql()) {
-            // Encode any 4-byte UTF-8 characters.
-            $value = StringHelper::encodeMb4($value);
-        }
-
-        return $value;
-    }
-
-    /**
-     * Parse ref tags in URLs, while preserving the original tag values in the URL fragments
-     * (e.g. `href="{entry:id:url}"` => `href="[entry-url]#entry:id:url"`)
-     *
-     * @param string $value
-     * @param ElementInterface|null $element
-     * @return string
-     */
-    private function _parseRefs(string $value, ElementInterface $element = null): string
-    {
-        if (!StringHelper::contains($value, '{')) {
-            return $value;
-        }
-
-        return preg_replace_callback('/(href=|src=)([\'"])(\{([\w\\\\]+\:\d+(?:@\d+)?\:(?:transform\:)?' . HandleValidator::$handlePattern . ')(?:\|\|[^\}]+)?\})(?:\?([^\'"#]*))?(#[^\'"#]+)?\2/', function($matches) use ($element) {
-            [$fullMatch, $attr, $q, $refTag, $ref, $query, $fragment] = array_pad($matches, 7, null);
-            $parsed = Craft::$app->getElements()->parseRefs($refTag, $element->siteId ?? null);
-            // If the ref tag couldn't be parsed, leave it alone
-            if ($parsed === $refTag) {
-                return $fullMatch;
-            }
-            if ($query) {
-                // Decode any HTML entities, e.g. &amp;
-                $query = Html::decode($query);
-                if (mb_strpos($parsed, $query) !== false) {
-                    $parsed = UrlHelper::urlWithParams($parsed, $query);
-                }
-            }
-            return $attr . $q . $parsed . ($fragment ?? '') . '#' . $ref . $q;
-        }, $value);
+        return parent::serializeValue($value, $element);
     }
 
     /**
@@ -965,62 +687,6 @@ class Field extends \craft\base\Field
     }
 
     /**
-     * Returns the available Redactor config options.
-     *
-     * @param string $dir The directory name within the config/ folder to look for config files
-     * @return array
-     */
-    private function _getCustomConfigOptions(string $dir): array
-    {
-        $options = ['' => Craft::t('redactor', 'Default')];
-        $path = Craft::$app->getPath()->getConfigPath() . DIRECTORY_SEPARATOR . $dir;
-
-        if (is_dir($path)) {
-            $files = FileHelper::findFiles($path, [
-                'only' => ['*.json'],
-                'recursive' => false
-            ]);
-
-            foreach ($files as $file) {
-                $filename = basename($file);
-                if ($filename !== 'Default.json') {
-                    $options[$filename] = pathinfo($file, PATHINFO_FILENAME);
-                }
-            }
-        }
-
-        ksort($options);
-
-        return $options;
-    }
-
-    /**
-     * Returns a JSON-decoded config, if it exists.
-     *
-     * @param string $dir The directory name within the config/ folder to look for the config file
-     * @param string|null $file The filename to load.
-     * @return array|false The config, or false if the file doesn't exist
-     */
-    private function _getConfig(string $dir, string $file = null)
-    {
-        if (!$file) {
-            $file = 'Default.json';
-        }
-
-        $path = Craft::$app->getPath()->getConfigPath() . DIRECTORY_SEPARATOR . $dir . DIRECTORY_SEPARATOR . $file;
-
-        if (!is_file($path)) {
-            if ($file !== 'Default.json') {
-                // Try again with Default
-                return $this->_getConfig($dir);
-            }
-            return false;
-        }
-
-        return Json::decode(file_get_contents($path));
-    }
-
-    /**
      * Returns the Redactor config used by this field.
      *
      * @return array
@@ -1030,7 +696,7 @@ class Field extends \craft\base\Field
         if ($this->configSelectionMode === 'manual') {
             $config = Json::decode($this->manualConfig);
         } else {
-            $config = $this->_getConfig('redactor', $this->redactorConfig) ?: [];
+            $config = $this->config('redactor', $this->redactorConfig) ?: [];
         }
 
         // Give plugins a chance to modify the Redactor config
@@ -1045,43 +711,19 @@ class Field extends \craft\base\Field
     }
 
     /**
-     * Returns the HTML Purifier config used by this field.
-     *
-     * @return HTMLPurifier_Config
+     * @inheritdoc
      */
-    private function _getPurifierConfig(): HTMLPurifier_Config
+    protected function defaultPurifierOptions(): array
     {
-        $purifierConfig = HTMLPurifier_Config::createDefault();
-        $purifierConfig->autoFinalize = false;
-
-        $config = $this->_getConfig('htmlpurifier', $this->purifierConfig) ?: [
-            'Attr.AllowedFrameTargets' => ['_blank'],
-            'Attr.EnableID' => true,
-            'HTML.AllowedComments' => ['pagebreak'],
-            'HTML.SafeIframe' => true,
-            'URI.SafeIframeRegexp' => '%^(https?:)?//(www\.youtube(-nocookie)?\.com/embed/|player\.vimeo\.com/video/)%',
-        ];
-
-        foreach ($config as $option => $value) {
-            $purifierConfig->set($option, $value);
-        }
-
-        // Give plugins a chance to modify the HTML Purifier config, or add new ones
-        $event = new ModifyPurifierConfigEvent([
-            'config' => $purifierConfig,
-        ]);
-
-        $this->trigger(self::EVENT_MODIFY_PURIFIER_CONFIG, $event);
-
-        return $event->config;
+        $options = parent::defaultPurifierOptions();
+        $options['HTML.AllowedComments'] = ['pagebreak'];
+        return $options;
     }
 
     /**
-     * Returns the allowed inline CSS styles, based on the plugins that are enabled.
-     *
-     * @return string[]
+     * @inheritdoc
      */
-    private function _allowedStyles(): array
+    protected function allowedStyles(): array
     {
         $styles = [];
         $plugins = array_flip($this->_getRedactorConfig()['plugins'] ?? []);
@@ -1099,5 +741,22 @@ class Field extends \craft\base\Field
             $styles['font-size'] = true;
         }
         return $styles;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function purifierConfig(): HTMLPurifier_Config
+    {
+        $purifierConfig = parent::purifierConfig();
+
+        // Give plugins a chance to modify the HTML Purifier config, or add new ones
+        $event = new ModifyPurifierConfigEvent([
+            'config' => $purifierConfig,
+        ]);
+
+        $this->trigger(self::EVENT_MODIFY_PURIFIER_CONFIG, $event);
+
+        return $event->config;
     }
 }